refact(decman-lib): move ActionType with codec methods
The closed Daml unions map to a Rust enum with methods, so the enum
moves whole: both encode forms, both decode dispatches, and validate.
The two production panics become Encode errors. Integrators now parse
vault and self confirmations with one call. decman re-exports the
enum, so the HTTP JSON is unchanged by construction.
