+++ crates/common/src/api.rs
@@ -763,7 +763,7 @@ pub struct GrantRightsResponse {
 // ============================================================================
 
 /// Instrument identifier (admin + id)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
 #[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
 pub struct InstrumentId {
@@ -772,7 +772,7 @@ pub struct InstrumentId {
 }
 
 /// Credential claim (subject, property, value)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
 #[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
 pub struct Claim {
