+++ crates/decman-lib/src/catalog/mod.rs
@@ -1,3 +1,4 @@
 //! DecMan's protocol content.
 
+pub mod action;
 pub mod types;
