+++ crates/decman/src/server/types.rs
@@ -326,154 +326,11 @@ pub type OnboardingResponse = WorkflowResponse;
 // `FarConfig` now live in `decman_lib::catalog::types` and are re-exported
 // above so existing `crate::server::types::X` paths keep resolving.
 
-/// Structured action types for Vault governance
-#[derive(Clone, Debug, Serialize, Deserialize, utoipa::ToSchema)]
-#[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum ActionType {
-    // Governance (4)
-    GovernanceAddMember {
-        member: CantonId,
-        new_threshold: i64,
-    },
-    GovernanceRemoveMember {
-        member: CantonId,
-        new_threshold: i64,
-    },
-    GovernanceSetThreshold {
-        new_threshold: i64,
-    },
-    GovernanceSetTimeout {
-        new_timeout_microseconds: i64,
-    },
-    GovernanceAddAdditionalProposer {
-        additional_proposer: CantonId,
-    },
-    GovernanceRemoveAdditionalProposer {
-        additional_proposer: CantonId,
-    },
-
-    // Vault Deployment (2)
-    VaultDeployment {
-        vault_rules_cid: String,
-        vault_name: String,
-        share_symbol: String,
-        asset_instrument_id: InstrumentId,
-        limits: VaultLimits,
-        vault_backend_signatory: CantonId,
-        #[serde(default)]
-        vault_far_config: Option<FarConfig>,
-        allocation_factory_cid: String,
-        registrar_service_cid: String,
-    },
-    YieldEpochDeployment {
-        vault_rules_cid: String,
-        vault_cid: String,
-        asset_instrument_id: InstrumentId,
-        vault_backend_signatory: CantonId,
-    },
-
-    // Vault Operations (5)
-    VaultPause {
-        vault_id: String,
-    },
-    VaultUnpause {
-        vault_id: String,
-    },
-    VaultUpdateLimits {
-        vault_id: String,
-        new_limits: VaultLimits,
-    },
-    VaultUpdateBackend {
-        vault_id: String,
-        new_backend_signatory: CantonId,
-    },
-    VaultUpdateFarBeneficiaries {
-        vault_id: String,
-        new_beneficiaries: Vec<AppRewardBeneficiary>,
-    },
-
-    // Processor (1)
-    ProcessorDeploymentRequest {
-        vault_processor_rules_cid: String,
-        vault_backend_signatory: CantonId,
-        allocation_factory_cid: String,
-        #[serde(default)]
-        processor_far_config: Option<FarConfig>,
-        initial_supported_vaults: Vec<String>,
-    },
-
-    // Utility Onboarding (4)
-    UtilityCreateProviderRequest {
-        operator: CantonId,
-    },
-    UtilityCreateUserRequest {
-        operator: CantonId,
-    },
-    UtilitySetup {
-        operator: CantonId,
-        provider_service_cid: String,
-        user_service_cid: String,
-    },
-    UtilityAcceptHolderServiceRequest {
-        operator: CantonId,
-        provider_service_cid: String,
-        holder_service_request_cid: String,
-        holder: CantonId,
-    },
-    // Credential Actions (2)
-    CredentialOfferFree {
-        operator: CantonId,
-        user_service_cid: String,
-        holder: CantonId,
-        id: String,
-        description: String,
-        claims: Vec<Claim>,
-    },
-    CredentialAcceptFree {
-        operator: CantonId,
-        user_service_cid: String,
-        credential_offer_cid: String,
-    },
-
-    // DevNet (1)
-    DevNetFeatureApp {
-        amulet_rules_cid: String,
-    },
-}
-
-impl ActionType {
-    /// Validate the action's fields. Returns an error message if invalid.
-    ///
-    /// Catches obviously-malformed inputs (negative thresholds, non-positive
-    /// timeouts) before they reach Canton's Daml checks. Canton rejects bad
-    /// values too, but here we surface a clear 400 rather than a generic
-    /// submission error after the proposal contract is already on the wire.
-    pub fn validate(&self) -> Result<(), String> {
-        match self {
-            ActionType::GovernanceAddMember { new_threshold, .. }
-            | ActionType::GovernanceRemoveMember { new_threshold, .. }
-            | ActionType::GovernanceSetThreshold { new_threshold } => {
-                validate_threshold(*new_threshold).map_err(|e| e.to_string())
-            }
-            ActionType::GovernanceSetTimeout {
-                new_timeout_microseconds,
-            } => validate_timeout(*new_timeout_microseconds).map_err(|e| e.to_string()),
-            ActionType::VaultDeployment {
-                vault_far_config: Some(far),
-                ..
-            }
-            | ActionType::ProcessorDeploymentRequest {
-                processor_far_config: Some(far),
-                ..
-            } => validate_beneficiary_weights(&far.beneficiaries).map_err(|e| e.to_string()),
-            ActionType::VaultUpdateFarBeneficiaries {
-                new_beneficiaries, ..
-            } => validate_beneficiary_weights(new_beneficiaries).map_err(|e| e.to_string()),
-            _ => Ok(()),
-        }
-    }
-}
+/// `ActionType` — including its Daml `Value` codec (`to_vault_proto` /
+/// `from_vault_proto` / `to_self_proto` / `from_self_proto`) and `validate`
+/// — now lives in `decman_lib::catalog::action`, re-exported here so
+/// existing `crate::server::types::ActionType` paths keep resolving.
+pub use decman_lib::catalog::action::ActionType;
 
 // The protocol validators (`validate_threshold`, `validate_timeout`,
 // `validate_unique_issuers`, `validate_self_issued_requirements_have_claims`,
